Show character inventory on shop screen [Re: nmkataoka/adv-life#69]
Shop screen now shows the shop's inventory on the left and the player's
inventory on the right side. Refactored the character inventory
screen so I could reuse the inventory display.
